move assets, countryCode, and intlData into constants folder and modify its usage
