+++ app/constants/assets.js
